Replace reductions in LULESH
From looking at comm counts, I found that the non-scalable
communication was due to the reductions.

Three of the reductions were being used to check for error conditions.
Brad noted that these could all be replace with foralls+halt.

While I was looking at reductions, replaced a minloc reduce with a
plain ol' min reduce, since it just dropped the loc value on the
floor.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@20762 3a8e244f-b0f2-452b-bcba-4c88e055c3ca