Initial support for .localAccess accessor
This adds initial support for a .localAccess method to arrays in order
to optimize for the case when an array access is known to be local to
the current task's locale.  At this point, I have only implemented it
for DefaultRectangular (which is essentially a normal access) and
BlockDist (which forwards to DefaultRectangular and relies on normal
bounds checking to keep things honest when runtime checking is on).

using localAccess has shown to improve lulesh's scalar code by about
30% in some of my experiments, so a next commit will be a version of
lulesh that makes use of localAccess.

Note that, at present, we haven't decided whether or how to expose
this to the user, so are considering it an undocumented feature until
we get more experience with it, determine whether it's the best way
to get these benefits, etc.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@20766 3a8e244f-b0f2-452b-bcba-4c88e055c3ca