Remove contention over a sync variable in the psReduce function in HPL
In the psReduce function that is called within LUFactorize, there was a sync
var used to lock another variable (containing the maximum value seen so far) on
every iteration.  Instead of locking it in every iteration, lock it only when
the local max is greater than the approximate global max. After obtaining the
lock, recheck for the true global max.

This takes each psReduce call from about 0.02 seconds to 0.001.

[Reviewed by Vass]


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@20770 3a8e244f-b0f2-452b-bcba-4c88e055c3ca