Minor LULESH edits
This is a set of edits from my local lulesh directories that are designed
to have minor performance/code clarity impacts, and to bring these two
copies of the code into greater synch with one another (to make diff-ing
them simpler).  Ultimately, the sparse vs. dense material elements work
in the release version should be merged back into my study version as well
to minimize those differences, but I didn't take that on here because it
breaks my use of 'local' statements until we get sparse distributions
committed to trunk.

Changes to both versions:
-------------------------
* removed the redundant zeroing out of the massAccum array, as
  noted by Sung

* replaced some curly brackets with 'do' for single-line loops

* other minor code cleanups to minimize diffs


Changes to the release version:
-------------------------------
* moved dtfixed into a config const (the MPI version apparently
  changes this value occasionally)

* promoted the ability to time each iteration (and get an average
  time/iteration value) to the release version

* param-unrolled the loop in localizeNeighborNodes() which seemed
  to have benefit in the studies directory


Changes to the studies version:
-------------------------------
* unified comments and formatting at global scope

* unified var->const formatting changes and tabbing with release
  version

* changed out/inout intents to ref, to match the release version

* added some comments to note future work





git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@20772 3a8e244f-b0f2-452b-bcba-4c88e055c3ca