Adding -snoRefCount to my study lulesh
This adds -snoRefCount to my study lulesh, which seems to
improve things by about 10% for the current local version
that I'm using.  Note that this probably shouldn't (or
couldn't) be used with the release version because one of
the main differences between the two is that in the study
version, all arrays are hoisted to the global scope to
avoid creation/destructrion overhead (so leaking the local
arrays in the release version would probably run us out of
memory quickly).



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@20774 3a8e244f-b0f2-452b-bcba-4c88e055c3ca