localAccess version of lulesh
This is a version of lulesh that uses the new prototype localAccess
method to make accesses known to be local cheaper.  I see
approximately a 30% improvement in execution time for a 15x15x15
problem size with this change.  More opportunities for applying
these localAccesses may still exist, I need to look more.

Note that I'm not promoting this to the release version because it
makes the code look significantly uglier and the vast majority of
cases that I transofmred ought to be handled by a compiler
optimization that links the leader domain of the loop as being the
source of the index variable used for an access to an array over that
same domain.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@20775 3a8e244f-b0f2-452b-bcba-4c88e055c3ca