Reduce regressing test count from 39 to 28.
Language change? Formerly, one could declare a variable with an explicit type
that was generic.  But then if the initializer has a concrete type all is
forgiven.  I decided to defer this issue by rewriting tests in which this
language nicety did not seem to be the main point.
    test/exercises/boundedBuffer.chpl
    test/classes/deitz/test_type_in_class4.chpl
    test/classes/deitz/test_type_in_class2.chpl
    test/classes/deitz/test_type_in_class3.chpl
    test/classes/deitz/test_default_param.chpl

This is fallout from changing the type of a localeID internally.  Casting loc to
the index type is an easy fix.
    test/optimizations/sungeun/RADOpt/access.chpl

Seemingly, we are accessing fewer locale expressions.  Good, right?
    test/multilocale/deitz/oneOrMoreLocales/test_dot_locale_error.good

Updated to obtain the Locales array through rootLocale.
    test/types/tuple/bradc/commTuple.chpl

Now, one can create a new locale.
    test/types/locale/bradc/createNewLocale.good

setLocale() does not work before _here is established for each node, so the
Locales array must be populated explicitly.
    modules/internal/ChapelRootLocale.chpl

Found and removed dead code for setting the parent class of an iterator class.
This is handled by the new call to addRootType().
    compiler/AST/iterator.cpp

Added some tests to flag the addition of duplicate children.
    compiler/passes/scopeResolve.cpp
    compiler/resolution/generics.cpp
    compiler/resolution/functionResolution.cpp

The test for deleting a record has moved into functionResolution.cpp.  Since the
call to chpl_here_free() is resolved before running checkResolved(), the error
catch there is now too late.
    compiler/passes/semanticChecks.cpp
    compiler/resolution/functionResolution.cpp

An on block cannot look up _here.getChild(sublocID) before here is established.
So we create a special case for "on" clauses that appear in the RootLocale module.
    compiler/passes/cleanup.cpp

In iterator classes remove the object parent if it has acquired another parent
after protoIterator.  There's probably a less kludgy way of doing this, like
checking that the iterator class has a parent after it is constructed.... 
    compiler/resolution/functionResolution.cpp


git-svn-id: http://svn.code.sf.net/p/chapel/code/branches/collaborations/caseyb@20788 3a8e244f-b0f2-452b-bcba-4c88e055c3ca