Replaced `pwd` with $cwd in start_test.
start_test currently ascends the directory tree in pursuit of .skipif files
until the current directory == $basedir.

Since $basedir is set using $cwd, the current directory in the above
test should be determined based on $cwd, rather than `pwd`. Otherwise
in the presence of symbolic links the equality may not hold even when
the current directory is, in fact, $basedir.

$cwd to `pwd` in csh seems the same as `pwd` to `pwd -P` in bash.

I did not put quotes around $cwd because throughout start_test
we consistently rely on paths/file names being shell-kosher.

There is another use of `pwd` in start_test, but that one seems harmless,
so I left it alone.

While here, I made it print the ancestor filename that caused the skip.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@20791 3a8e244f-b0f2-452b-bcba-4c88e055c3ca