Bring failure count down to 25.
Rewrite the caseyb test to match updated ChapelRootLocale and ChapelLocale
interfaces.
    test/users/caseyb/BaseLocale.chpl
    test/users/caseyb/test.chpl
    test/users/caseyb/test.good
    test/users/caseyb/SubLocale.chpl

I thought I already fixed this....
    test/optimizations/sungeun/RADOpt/access.chpl

locale.alloc() changed back to a form which takes a byte count.  This error is
masked in other tests by the recursive iterator args copy function, but exists
nonetheless.  It was necessary to get caseyb/test.chpl to run.
    modules/internal/ChapelLocale.chpl
    compiler/AST/primitive.cpp
    compiler/AST/expr.cpp
    compiler/passes/parallel.cpp
    compiler/resolution/wrappers.cpp
    compiler/resolution/lowerIterators.cpp
 


git-svn-id: http://svn.code.sf.net/p/chapel/code/branches/collaborations/caseyb@20796 3a8e244f-b0f2-452b-bcba-4c88e055c3ca