Add comment with function name in generated code for runtime calls that use the ftable
[reviewed by diten]

Michael's codegen change removed the commented function name that
followed the ftable index in the generated code for runtime calls to
the tasking and comm layers (used for begin, coforall, cobegin, and on
statements).

This change puts the function name at the beginning of the line with
the runtime call.  I don't think it's as clear as it was previously
since it's not right next to the ftable index, but this is the
cleanest way to do it for now without changing the interface to
codegenCall().



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@20800 3a8e244f-b0f2-452b-bcba-4c88e055c3ca