Removing --fast flags from correctness tests
I'm removing the --fast flags from the SSCA2 correctness tests because

(a) using -O3 with the PGI compiler results in nondeterministic timeouts
(in spite of the fact that the timeout is set to 30 minutes!) and

(b) turning off the runtime checks makes it easier for bugs to creep
into the code.  Specifically, the 3rd compilation option results in an
out-of-bounds array access when --fast is removed that we should have
caught before this point.  I hope to fix this bug during this triage
rotation, but feel that it shouldn't hold up this commit since (i) the
bug exists with or without the --fast flag and (ii) having a consistent
failure beats having a nondeterministic timeout.


More generally, this is making me think that we should try to avoid
using --fast in correctness tests and use specific sub-flags (like
flipping on -O2 or -O3) if the goal is to speed up the execution time
(at the cost of greater compilation time).  In this case, my thought
is that compilation time is the bottleneck for the correctness tests,
not execution time (but please correct me if I'm wrong).




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@20802 3a8e244f-b0f2-452b-bcba-4c88e055c3ca