Match a gflops value printed in exponential notation.
Sometimes the gflops rating is so small that it ends up being printed in
exponential notation.  (And perhaps someday it will be so large that the same
happens!)  Adjust the regular expression that matches the printed gflops value
so that it works whether or not exponential notation is used.

While here, I also got rid of the superfluous "g" suffix in both sed scripts.
This is only needed in order to have the script work more than once per input
line, which cannot happen for this test.

Also while here, I fixed the comment before the sed commands so that it says
what is actually being done.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@20803 3a8e244f-b0f2-452b-bcba-4c88e055c3ca