Fixing arguably non-portable task parallel spec example
Qthreads has failed non-deterministically on the syncCounter example
in this chapter for reasons related to assumptions about task
scheduling, busy-waiting, and starvation (see chapel-developers thread
called "The Meaning of Begin" -- Sept 2012 -- for far more discussion
on this topic).

As we were wrestling with these questions, it was pointed out that, at
the very least, the use of begins and busy-waiting in this example was
quite orthogonal to what the example was written to illustrate and
that using a cobegin would achieve the same goal without raising the
semantic questions.  I invited someone to submit a counterproposal to
the example, but either never got one, or else it got lost in my inbox
(very likely -- that conversation came at a busy time).  In either
case, the nondeterministic failure has continued to be noisy and is
simple to fix.

In doing this fix, I also rewrote some of the text around the example
in an attempt to clarify it -- hopefully successfully.

Note that I have not taken on the bigger task of updating the spec to
clarify the semantics of begin (something that we did receive proposed
text for, but I lost track of whether or not we ever came to a clear
consensus and haven't gone back to it yet... (this being my triage
week, my goal is to stamp out noise rather than doing other benevolent
improvements... :)



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@20806 3a8e244f-b0f2-452b-bcba-4c88e055c3ca