Make heap allocation a little less conservative
[reviewed by diten]

This commit makes the selection of data being heap allocated a little
less conservative.

During the flatten functions pass, we were being too conservative
about what outer vars got passed by reference to the nested functions
we flattened.  In some cases (many cases in SSCA2), this resulted
putting non-var scalar coforall index vars on the heap.  Now we only
pass the vars by ref if the variable is not a coforall index *or* it's
parent function is not var.  This is still conservative, and there is
a comment in the relevant places in the code noting this.  Note also,
with further analysis, we could avoid passing by reference any scalar
var that is only *used* in the nested functions and that might improve
the back end C compiler analysis (or not).

During the heap allocations pass, all heap allocated variables used
within an on/begin/cobegin/coforall were not being freed.  Now we only
avoid freeing heap vars that are used in a begin or non-blocking on
statement or if they are coforall index vars (see below about coforall
and "vector fork").

We were also placing stack variables on the heap if they were being
used in an on statement since some communication layers require
communication only between heap segments.  We have two configurations
that don't have this requirement (CHPL_COMM=ugni, CHPL_COMM=gasnet
with CHPL_GASNET_SEGMENT=everything), so I added logic to avoid
putting such vars on the heap for these configurations.

Related changes also in this commit:

- Made CHPL_GASNET_SEGMENT and CHPL_COMM_SUBSTRATE available at
  compile time as per above.  CHPL_COMM_SUBSTRATE is currently not
  needed for this change, but I added it anyways since it seems like
  something we might eventually use and it seemed weird to only have
  CHPL_GASNET_SEGMENT without CHPL_COMM_SUBSTRATE.

- Added a flag called FLAG_COFORALL_INDEX_VAR and used that instead of
  FLAG_HEAP_ALLOCATE since that is really why the var was tagged with
  FLAG_HEAP_ALLOCATE.  I left FLAG_HEAP_ALLOCATE in there for now even
  though it's not used for anything else.

This change was motivated by SSCA2. Depending on the problem size,
nearly half of the leaked memory was due to heap allocated data that
should have been freed or (to a lesser extent) did not need to be put
on the heap in the first place.  Another known source of unnecessarily
heap allocated memory leaks are the coforall args.  The problem is
that we allocate the bundle for the args in a loop and those bundles
cannot be freed until the task using each bundle starts.  The current
plan is to deal with this using a "vector fork", where we have a
single allocation for all the bundles and pass that all to a new
runtime routine that does all the tasks starts in a single call.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@20807 3a8e244f-b0f2-452b-bcba-4c88e055c3ca