Added support for LCG random number generator for RA
The current Random number generator for RA was (purposefully) designed
to not spread updates out evenly, but has the downside that, when run
at smaller problem sizes (as we do for correctness testing), raises
the chances that conflicting updates will occur and that verification
will fail (at least, this is our current theory for the relatively
frequent failures in the non-atomic versions).

This commit adds a config param to RARandomStream.chpl that permits
one to select between an LCG algorithm that was added as an option to
the official HPCC RA after the first few years of the competition or
the original algorithm.  The LCG is used by default to give a balanced
computation for the purposes of testing.  It may be that for
performance runs, the unbalanced case would continue to be considered
more interesting or challenging for "real" users (I'm not sure, just
speculating).

[It also occurred to me while walking home and thinking about this
that we could probably run with a large table size but small number of
updates using the original algorithm to dial down the probability of
conflicts.  But since the reference version supports two random number
streams, it seems nice for us to support both as well.  And we do it
so much more nicely than their "duplicate all source files and
funtions" approach -- bleah].

Updated existing tests that relied on the old behavior, typically by
adding -suseLCG=false to their compopts file.  In cases that seemed
interesting, I also added a second compopts line that tests the LCG
algorithm.

Perhaps more importantly, I added a new test -- testRAStream.chpl that
simply generates the random stream in parallel and prints it out.
Then I created C codes that did the same using the pseudocode in the
reference version and used them to create the .good files (this could
also be done at test time to reduce the file sizes stored in SVN but I
was too lazy to worry about a portable implementation of that tonight.
The best approach I came up with was to call out to the C routines
from within Chapel, piping the output to the .good file directly to
avoid having to make assumptions about compilers and flags and such).

I didn't do the full indenting of the RARandomStream.chpl file
because I wanted to keep the diff minimal, but will do that in
a separate checkin.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@20808 3a8e244f-b0f2-452b-bcba-4c88e055c3ca