Adding a few tests I forgot to svn add in the previous commit
Two are new .good files for the tests that I duplicated the .compopts
on.

The third is a .numlocales file for my new stream generation test,
just to make the GASNet case more interesting.

Sadly, my "check in indentation changes separately" plan failed
because I thought once the svn commit had gotten past a file,
it was OK to edit it, but it seems I was too fast and the
indentation change went into the original commit.  Oh well.

Finally, it occurs to me: I used a config param for useLCG
thinking it would be perforamance critical, but given that
it shows up in the follower iterator... it may not be such
a big deal (and would save a few extra compiles in testing
if converted to a config const).  I don't consider it a big
enough deal to bother with for now, but wanted to note it
while it was on my mind.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@20809 3a8e244f-b0f2-452b-bcba-4c88e055c3ca