Dialing down problem size of this test
When I was putting this test together, I tested it for a million
iterations to gain confidence that there weren't any funny numeric
artifacts that would happen somewhere deep in the iteration and then,
due to fatigue, checked it in that way.  In retrospect, that's
ridiculously long (in terms of size of .good file and execution time),
and, having verified it's correct for that many iterations, running it
for 1000 for the test seems far more reasonable.  This commit does
that.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@20812 3a8e244f-b0f2-452b-bcba-4c88e055c3ca