Fix bug due to r20807
The heap allocation fix of r20807 was a little to agressive when
deciding which heap vars to *not* free (all var being pass to
functions in the functions in fnsContainingTaskll should *not* be
freed).  This resulted in a baseline regression of the test
test/parallel/begin/deitz/test_nested_fn_begin1.chpl.  As it turned
out, the bug was being masked in non-baseline testing by remote value
forwarding because the value was only being read.

Added a test that would have always failed under valgrind without this
bug fix.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@20815 3a8e244f-b0f2-452b-bcba-4c88e055c3ca