Changing RA correctness execution options
In this commit, I'm dialing up the RA problem size used for
correctness checking, but keeping the number of updates computed the
same as before.  My hope is that the execution time will be similar to
what we had before (modulo operations that are O(table size)), and
that the updates should be spread out over a larger table size, so
less likely to conflict.

Prior to this change, we were running with a table size of 1024, and
it seemed that at that size, the chances of getting too many conflicts
was high enough that we'd get failures regularly, even with the LCG
random number generator.  My hope is that these failures will go away
with this new setting.  If not, further investigation may be
required...



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@20816 3a8e244f-b0f2-452b-bcba-4c88e055c3ca