Removing the nodistrib subdirectory of hpcc/
I was going to update the copy of ra in this directory as I did the
others yesterday because it failed last night in a similar way; but
I've generally been wondering about the proliferation of versions of
things and wondering what could be pruned, so started by doing some
detective work on why this directory exists and whether it should
continue to do so.  My verdict (supported by Sung) is that it should
be removed.

Historically, this directory existed to create distribution-free
versions of the HPCC benchmarks -- copies in which no distribution was
used in order to maximize performance for single-locale runs (by
avoiding degenerate implementations of Block, for example).  This
was done primarily for XMT performance work (a single-locale machine).

Since that time, we've supported similar versions not by cloning
benchmarks, but by adding config params to our benchmarks that say
whether or not to target multiple locales.  LULESH and SSCA2 are
examples of this.  So if we want to have single-locale versions of
these codes, we should beef up the official versions (a 1-3 line
change typically) rather than cloning them in this way.

That said, I'm not undertaking that change with this commit because
it's been a long time since anyone wanted a single-locale version of
these to study that I'm aware of.  So in the interest of time, let's
cross that bridge when we come to it now that we're more cognizant, as
a group, that it can be done.




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@20817 3a8e244f-b0f2-452b-bcba-4c88e055c3ca