Fix to regression relating to array type aliases
As a result of the chapel.ypp refactoring of r20223 to introduce array
literals, the treatment of array types in 'type' alias statements was
inadvertantly changed in a way that broke them.  Presumably, we either
didn't have tests filed that triggered this case, or they slipped
through the cracks in the regression testing that week due to other
higher-profile failures.

This fix is somewhat unsatisfactory in that I can't really explain
precisely what was broken about the code prior to it, nor what about
this fix makes it work again.  It was more a case of noticing that the
test had worked in the past, finding the revision that broke it,
noting the different IR we were generating as a result of that change,
and reverting it to the original IR.  In that sense, this change
treats the compiler as a black box artifact and relies on the wisdom
of the ancients rather than being a change I can explain.  But I think
it's worthwhile to proceed with in the interest of fixing it quickly,
and using the rationale that things are no worse/different than they
would have been if r20223 hadn't inadvertantly changed the IR...

I debugged it by looking at differences in the generated code, tracing
them back to the type alias statement, and looking at the r20222
vs. r20223 parser rules for array type aliases.  In r20222, they were
turned into forall expressions (buildForallLoopExpr) while array types
were turned into array types (new CallExpr("chpl__buildArrayRuntimeType", 
...)).  In r20223, these two rules were unified into one and so both
were handled by building an array runtime type which seems to be the
thing that broke array type aliases.

In this commit, what I'm doing is to keep the unified rule and have it
build the array type, but then if I end up in the type alias rule, I'm
plucking out the domain and element type expressions and calling the
same forall expression that r20222 would have out of them.  This seems
to fix the bug.

Another implementation I considered but rejected was to create a little
struct to store the index, domain, and element expressions in the
'array_type' rule, to pass that upward, and have the containing rule
either build a runtime array type or forall expression depending on
the context; the trouble with this approach was that it didn't support
recursive array types well -- for that you'd need more of a recursive
type like an Expression, and at that point, it felt very similar to
what buildArrayRuntimeType was already doing.

It occurs to me as I write this that another way to have addressed
this might have been to unify the array type alias case with the
square bracket subset of the 'for_expr' rule.  That may have been
another way to recapture what was going on in r20222.  That said, I
think it makes more logical sense to treat array types consistently in
the parser which favors the current approach (though given that we
view array types as a form of forall expression, perhaps all three
cases should really be unified).

I feel extremely non-confident that what I've done here is very
general w.r.t. things like creating type aliases for arrays of arrays
or using index variables within array type aliases (skyline array
aliases, which we don't support in a non-type-alias format yet
anyway); but I'm not convinced those features ever worked in the first
place.  I am checking in one future that shows that type aliases for
arrays-of-arrays doesn't work, and will check after this commit to see
whether it did in r20222.  In any case, more complex array type
aliases will require additional work.

Finally, I noticed what look to be some irregularities in the
'array_type' rule and noted them in a comment in the sources, though I
didn't try to undertake fixing them.  I would expect that we'll run
into these in the future (either in the form of syntax errors or
overly permissive array types) and can deal with them then.




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@20820 3a8e244f-b0f2-452b-bcba-4c88e055c3ca