Re-enable support for type-defs of arrays of arrays
[reviewed by Tom; and I forgot to say in 20820 that John reviewed it]

This is an extension to r20820.  After committing that fix, I went
back to see whether the arrOfArrTypedef.chpl future that I checked in
at the end of the bugfix had worked with r20222.  It essentially makes
sure that arrays-of-arrays can be typedef'd:

      type ArrOfArr = [1..2] [1..3] real;

It turns out that it did, so I felt compelled to make it work again.

r20820 was essentially a change to convert a "build array runtime
type" call into a forall expr, to restore the behavior of r20222.
However, it was shallow, only converting the outermost "build array
runtime type" function and ignoring any that were nested within.

The fix to make arrays of arrays work was simply to not be so lazy and
look into the element type of the array to see whether it too was an
array (i.e., a call to "build array runtime type").  If so, we
recursively insert forall exprs (as -- not surprisingly, in retrospect
-- the r20222 parser did).  

To keep the parser cleaner, I moved this logic into a new helper
function in build.cpp and also made the logic a bit more bullet-proof
in terms of making sure it gets the CallExpr form that it expects.

Now that all three of my array alias tests are working, I merged them
into a single test, just to reduce the time spent compiling.  I'm
still somewhat amazed that we had no pre-existing tests of either of
these typedef forms...



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@20823 3a8e244f-b0f2-452b-bcba-4c88e055c3ca