Add future test case for --no-inline resulting in incorrect code
This code is an example of --no-inline generating incorrect code.
From the future:

     When compiled with --no-inline (and run under valgrind), this
     test shows an example of the promotion wrapper combined with the
     coerce wrapper returning a the address of a record that is on the
     stack.  I'm not exactly sure which part is to blame, but when
     both functions are inlined the data gets put on the heap as it
     should be.

Unfortunately, this future will never be run because it only fails
under valgrind and we don't test futures under valgrind testing.
Getting a simple test that always fails is difficult since we're
misusing the stack, and I didn't want to add more intermittent
failures to the testing.

FYI, I found this while looking at the intermittent baseline testing
failures.  Brad says these are relatively recent, and this makes sense
because they are due to the lazy RAD optimization implemented last
summer.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@20827 3a8e244f-b0f2-452b-bcba-4c88e055c3ca