Added Chapel-level param functions isRecordType(type t) and chpl__isDmap(x).
[Reviewed by David.]

isRecordType tests whether the argument type is a Chapel-level record.
Therefore it needs to return false for other Chapel types that happen
to be implemented with records. It also needs to return false for
records that are treated by the compiler specially (which probably is
the same as the previous criterion).

The only missing piece is to make it return 'false' on atomic types.
This can be done by providing an isAtomicType() function. The latter
could be implemented e.g. like _isUnsignedType(), by comparison
of the argument against all valid atomic types.

Also adding a test to check that isRecordType, isClassType, and
isUnionType work correctly.

Also adding type-tests.todo.future. As pieces of it start working,
let's migrate the corresponding code into type-tests.chpl.
Once everything is working, fix up type-tests.chpl and remove
type-tests.todo.*.

chpl__isDmap() tests whether its argument is the result
of 'new dmap(...)'. This was needed for isRecordType().


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@20831 3a8e244f-b0f2-452b-bcba-4c88e055c3ca