Report an error when declaring an associative domain with an unsupported index type.
[Reviewed by David.]

Whether it is legal to create an associative domain with a given idxType
is determined by the domain's domain map. So this check is specific
to default associative domains. In which case the legality is determined
by the availability of chpl__defaultHash().

Enumerated, opaque, and sparse domains with the default domain map are
handled separately and so are considered here.

Default associative domains over tuples and records also seem work.
I allowed those even though I don't know why they work.
Feel free to adjust this.

The test assoc-idxtype-error.chpl tests that the error is/is not raised.
Additions to the set of legal cases and
improvements to the error message are welcome.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@20832 3a8e244f-b0f2-452b-bcba-4c88e055c3ca