This patch fixes three bugs in QIO
The actual bug fixes are in qio.c and qio.h; the rest is cleaning
up and adding testing code that reveals the problem.

The first bug is that if QIO_HINT_PARALLEL was given when opening
a new (or zero length) file, we'd try to mmap 0 bytes
(always a bad idea), so now we don't do that.

The second was in the read path for reading bits at a time; the
bits I/O code uses ch->cached_end_bits and zeros out ch->cached_end,
so that if we e.g. write 8 bits (which would just store it in our
word-accumulator) and then write using one of the byte-aligned
normal I/O calls, we appropriately advance 1 byte forward...
The bug was that when the read bits call needed to do regular
I/O, it did not put ch->cached_end back to normal (and so caused
an assert failure in certain scenarios especially with
QIO_HINT_PARALLEL).

The third was that after writing some bits to a channel, the
offset reported by qio_channel_offset would be wrong
(e.g. writing 8 bits and then calling qio_channel_offset would return 0
 and not 1). Fix that by just making qio_channel_offset take
bit buffer usage into account. After this patch, qio_channel_offset
will always return the byte offset at which the next byte
would be read/written if you called qio_channel_read_int8.
In other words, if we wrote 5 bits, it would round that
up to a byte and return an offset of 1 (since when writing
normally after writing with bits, we pat 0 bits until we
get to a byte boundary).

[OK'd by Brad]



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@20834 3a8e244f-b0f2-452b-bcba-4c88e055c3ca