Tighten up line number assignment to AST nodes in the compiler.
[Reviewed by Sung.]

Background: an AST node is associated, via its 'astloc' field, with
some location within the Chapel program being compiled, i.e. a file
name and a line number. The 'astloc' field is set when the node is
created, in the BaseAST::BaseAST constructor, from the global variable
'currentAstLoc'. Except: during parsing, yyfilename and yystartlineno
are consulted instead. The 'astloc' field may be modified after the node
is created, explicitly or via reset_ast_loc(), but often it stays the same.
    'currentAstLoc' is set mostly via a call to the SET_LINENO macro.
Prior to this patch, in many cases it was not set well, resulting in
AST nodes getting inappropriate location information. This patch
aims at improving this situation, including a means for partial checking
whether 'currentAstLoc' has been set consciously, and so perhaps correctly.

This commit may change the locations associated with some AST nodes.
However, the changes, if any, will be visible only while debugging
the compiler; the locations checked by our test suite have not changed.
Whenever you come across an AST node whose location you disagree with,
feel free to change it.


* How SET_LINENO works w.r.t. currentAstLoc has changed.

Code like this in the compiler:

    {
      some code;
      SET_LINENO(someAst);
      some more code;
    }

used to translate into

    {
      some code;
     currentAstLoc = someAst->astloc;
      some more code;
    }

Now 'currentAstLoc' is reverted to its previous value upon exiting
the scope where SET_LINENO appears, roughly like so:

    {
      some code;
     astlocT markAstLoc = currentAstLoc;
     currentAstLoc = someAst->astloc;
      some more code;
     currentAstLoc = markAstLoc;
    }

This is done using destructors (markAstLoc is actually a record,
whose destructor performs the restoration).


* Additional error checking is added.

When an AST node is created, in the BaseAST::BaseAST constructor,
a check is performed whether 'currentAstLoc' has been consciously set.
If not, the compiler aborts via INT_FATAL. This is to make sure
future code is always executed with currentAstLoc set at least to
something, somewhere up the call chain. This is of course only a partial
and approximate check of correctness of location info in AST.


* Additional SET_LINENO calls are inserted.

Those are mostly in the cases like the following, where there used to be
no SET_LINENO calls at all in the loop body:

  void someCompilerPass() {
    forv_Vec(FnSymbol, fn, gFnSymbols) {
      ...
      ... new someAstNode(...) ...
      ...
    }
  }

In some cases I created additional scopes (i.e. added { } pairs)
to make sure the effect of SET_LINENO is localized or when
multiple SET_LINENOs appeared in the same scope and so conflicted
for the local variable definition ('markAstLoc').

The latter is somewhat intentional - to raise awareness of a compiler
developer about multiple SET_LINENOs in the same scope. It is also
driven by the desire to leave existing SET_LINENOs as-is, without
having them declare new variables. I guess we could work around
the latter, if we wanted, by appending __LINE__ to the variable name.

In a couple of cases, I added a { } pair with a SET_LINENO for
just one of a couple of AST node creations. This was necessary
to avoid triggering the error above check. Alternative suggestions
are welcome.

* The "currentAstLoc has not been consciously set" condition
is now marked with currentAstLoc.filename==NULL. Ditto for
yyfilename/yystartlineno.

We used to use lineno==-1 to indicate both the above condition and the
condition "there is no good file/line for this AST, just use
<internal> (or somesuch)". But the error check added above requires
that we differentiate the two, so I used filename==NULL as the condition.
This is reflected in the BaseAST::BaseAST constructor.

We can go back and use lineno==-1 for the "not consciously set"
condition and lineno==0 for the other condition, but that will
require a little of a cleanup.


* Open questions:

 - Would be nice to have BaseAST::copy/copyInner copy the location info
   from the original. But this is not how it is done currently afaik
   and I wasn't bold enough to change it. So I had to wrap { } SET_LINENO
   around a couple of copy calls instead.

 - Would be nice to retire the calls to reset_ast_loc(). Again,
   this might change the current behavior and would not provide
   an absolute guarantee that all AST nodes in question will get
   the desired location info.


* Miscelanea:

- I believe my patch does not change line numbers during inlining
vs. the current behavior. That's because SET_LINENO acts the same in
this context (except perhaps some corners), and the line numbers are
reset (or not, if --preserve-inlined-line-numbers) using a separate
later call to reset_ast_loc().

- This change is intended to be an incremental improvement.  We also
discussed a desire to have a more manageable solution, e.g. to
use the line number of the statement you are inserting around (via
the insert* methods on block statements or expressions).
That is not addressed here.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@20835 3a8e244f-b0f2-452b-bcba-4c88e055c3ca